update by parai

+++ osek_test/GaInOS_Studio/Os_Cfg.h
@@ -37,7 +37,7 @@
 #define cfgOS_PRE_TASK_HOOK 0
 #define cfgOS_POST_TASK_HOOK 0
 
-/* For MC9S12(X) */
-typedef uint8_t OsCpuSrType;
-typedef uint8_t OsCpuIplType;
+/* For ARM */
+typedef uint32_t OsCpuSrType;
+typedef uint32_t OsCpuIplType;
 #endif /* _OS_CFG_H_ */
